+++ pio-pps/buf.h
@@ -13,6 +13,6 @@ struct cir_buf* cir_buf_init(uint32_t *buffer, int size);
 
 void cir_buf_push(struct cir_buf* buf, uint32_t value);
 
-float cir_buf_avg(struct cir_buf* buf);
+double cir_buf_avg(struct cir_buf* buf);
 
 void cir_buf_print(struct cir_buf* buf);
